# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RecordStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive"
				]
			},
			"AccountAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"remarks": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					}
				},
				"required": [
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldError": {
				"properties": {
					"value": {
						"type": "string",
						"nullable": true
					},
					"field": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"field",
					"type",
					"message"
				],
				"type": "object"
			},
			"EntityError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"fieldErrors": {
						"items": {
							"$ref": "#/components/schemas/FieldError"
						},
						"type": "array",
						"default": []
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileType": {
				"type": "string",
				"enum": [
					"stakeholder",
					"admin",
					"user"
				]
			},
			"AuthProfile": {
				"properties": {
					"remarks": {
						"type": "string"
					},
					"scopes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"mobileNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"type",
					"email",
					"username",
					"name",
					"id"
				],
				"type": "object"
			},
			"AuthResult": {
				"properties": {
					"token": {
						"type": "string"
					},
					"profile": {
						"$ref": "#/components/schemas/AuthProfile"
					}
				},
				"required": [
					"token",
					"profile"
				],
				"type": "object"
			},
			"RegisterProfile": {
				"properties": {
					"mobileNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"username",
					"name"
				],
				"type": "object"
			},
			"UpdateProfile": {
				"properties": {
					"remarks": {
						"type": "string",
						"nullable": true
					},
					"scopes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"mobileNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"type",
					"email",
					"name"
				],
				"type": "object"
			},
			"ProjectAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"estimatedCost": {
						"type": "number",
						"format": "double"
					},
					"remarks": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					}
				},
				"required": [
					"name",
					"code",
					"location",
					"estimatedCost",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProfileAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"scopes": {
						"type": "string"
					},
					"remarks": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"email",
					"type",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VoucherAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"series": {
						"type": "string"
					},
					"projectId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"totalCost": {
						"type": "number",
						"format": "double"
					},
					"remarks": {
						"type": "string"
					},
					"closed": {
						"type": "boolean"
					},
					"processedBy": {
						"type": "number",
						"format": "double"
					},
					"processedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					}
				},
				"required": [
					"series",
					"projectId",
					"description",
					"totalCost",
					"processedBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionType": {
				"type": "string",
				"enum": [
					"cash",
					"check"
				]
			},
			"TransactionAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"accountId": {
						"type": "number",
						"format": "double"
					},
					"projectId": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"transactionType": {
						"$ref": "#/components/schemas/TransactionType"
					},
					"details": {
						"type": "string"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"processedBy": {
						"type": "number",
						"format": "double"
					},
					"processedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					}
				},
				"required": [
					"accountId",
					"transactionType",
					"details",
					"amount",
					"processedBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DisburseProjectFund": {
				"properties": {
					"remarks": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkNumber": {
						"type": "string"
					},
					"transactionType": {
						"$ref": "#/components/schemas/TransactionType"
					},
					"series": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"processedBy": {
						"type": "number",
						"format": "double"
					},
					"accountId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"amount",
					"transactionType",
					"series",
					"description",
					"processedBy",
					"accountId"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "authorization"
			}
		}
	},
	"info": {
		"title": "foundation-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/account/getAll": {
			"get": {
				"operationId": "GetAllAccounts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AccountAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/account/createAccount": {
			"post": {
				"operationId": "CreateAccount",
				"responses": {
					"201": {
						"description": "created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountAttr"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AccountAttr"
							}
						}
					}
				}
			}
		},
		"/api/account/updateAccount/{id}": {
			"patch": {
				"operationId": "UpdateAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccountAttr"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					},
					"404": {
						"description": "not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AccountAttr"
							}
						}
					}
				}
			}
		},
		"/api/auth": {
			"post": {
				"operationId": "Auth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResult"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/profile/getAll": {
			"get": {
				"operationId": "GetAllProfiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AuthProfile"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/profile/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResult"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterProfile"
							}
						}
					}
				}
			}
		},
		"/api/profile/me": {
			"get": {
				"operationId": "Me",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthProfile"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
		"/api/profile/updateProfileStatus/{id}": {
			"patch": {
				"operationId": "UpdateProfileStatus",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordStatus"
						}
					}
				]
			}
		},
		"/api/profile/updateProfile/{id}": {
			"patch": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthProfile"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					},
					"404": {
						"description": "not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfile"
							}
						}
					}
				}
			}
		},
		"/api/profile/changePassword/{id}": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									},
									"currentPassword": {
										"type": "string"
									}
								},
								"required": [
									"newPassword",
									"currentPassword"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/profile/resetPassword": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/project/getAll": {
			"get": {
				"operationId": "GetAllProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProjectAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/project/getProjectVouchers/{id}": {
			"get": {
				"operationId": "GetProjectVouchers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/VoucherAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/project/getProjectTransactions/{id}": {
			"get": {
				"operationId": "GetProjectTransactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TransactionAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/project/createProject": {
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"201": {
						"description": "created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectAttr"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectAttr"
							}
						}
					}
				}
			}
		},
		"/api/project/updateProject/{id}": {
			"patch": {
				"operationId": "UpdateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectAttr"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					},
					"404": {
						"description": "not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectAttr"
							}
						}
					}
				}
			}
		},
		"/api/project/disburseProjectFund/{id}": {
			"post": {
				"operationId": "DisburseProjectFund",
				"responses": {
					"201": {
						"description": "created"
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DisburseProjectFund"
							}
						}
					}
				}
			}
		},
		"/api/project/depositProjectFund/{id}": {
			"post": {
				"operationId": "DepositProjectFund",
				"responses": {
					"201": {
						"description": "created"
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TransactionAttr"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}